+++ host-windows/build.rs
@@ -14,7 +14,7 @@ fn main() {
             r#"<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <assembly xmlns="urn:schemas-microsoft-com:asm.v1" manifestVersion="1.0">
   <assemblyIdentity
-    version="0.1.8.0"
+    version="0.1.9.0"
     processorArchitecture="amd64"
     name="FlexDisplay.Host"
     type="win32"/>